Feature/home14

+++ web/homework_12/src/repository/contacts.py
@@ -15,6 +15,19 @@ async def get_contacts(
         skip: int = 0,
         limit: int = 10
 ) -> List[Contact]:
+    """
+    Receive a list of contacts for a specific user
+    with specified pagination parameters.
+
+    Args:
+        db (Session): The database session.
+        user (User): The user to retrieve contacts for.
+        skip (int, optional): The number of contacts to skip. Defaults to 0.
+        limit (int, optional): The maximum number of contacts to return. Defaults to 10.
+
+    Returns:
+        List[Contact]: A list of contacts
+    """
     return db.query(Contact).filter(
         Contact.user_id == user.id
     ).offset(skip).limit(limit).all()
@@ -25,6 +38,17 @@ async def get_contact(
         user: User,
         contact_id: int
 ) -> Contact | None:
+    """
+    Get a specific contact for a specific user.
+
+    Args:
+        db (Session): The database session.
+        user (User): The user to get contact for.
+        contact_id (int): The ID of the contact to get.
+
+    Returns:
+        Contact | None: The contact, or None if the contact does not exist.
+    """
     return db.query(Contact).filter(
         and_(Contact.id == contact_id, Contact.user_id == user.id)
     ).first()
@@ -35,6 +59,20 @@ async def create_contact(
         user: User,
         contact: ContactUpdateSchema
 ) -> Contact:
+    """
+    Create a new contact for a specific user.
+
+    Args:
+        db (Session):  The database session
+        user (User): The user to create contact for.
+        contact (ContactUpdateSchema): The for the contact to create.
+
+    Raises:
+        e: Any exception that occurs during the database operation.
+
+    Returns:
+        Contact: The created contact.
+    """
     try:
         new_contact = Contact(**contact.dict())
         new_contact.user_id = user.id
@@ -54,20 +92,47 @@ async def update_contact(
         contact_id: int,
         contact: ContactUpdateSchema
 ) -> Contact | None:
-    db.query(Contact).filter(
-        and_(Contact.id == contact_id, Contact.user_id == user.id)
-    ).update(contact.dict())
+    """
+    Update a contact for a specific user.
+
+    Args:
+        db (Session): The database session.
+        user (User): The user to update contact for.
+        contact_id (int): The ID of the contact to update.
+        contact (ContactUpdateSchema): The updated data for the contact.
+
+    Returns:
+        Contact | None: The updated contact, or None if the contact does not exist.
+    """
+    origin = await get_contact(db, user, contact_id)
+    if origin is None:
+        return None
+    for key, value in contact.dict().items():
+        setattr(origin, key, value)
     db.commit()
 
-    return await get_contact(db, user, contact_id)
+    return origin
 
 
 async def delete_contact(
         db: Session,
         user: User,
         contact_id: int
-) -> Contact:
+) -> Contact | None:
+    """
+    Delete a contact for a specific user.
+
+    Args:
+        db (Session): The database session.
+        user (User): The user to delete contact for.
+        contact_id (int): The ID of the contact to delete.
+
+    Returns:
+        Contact | None: The deleted contact, or None if the contact does not exist.
+    """
     contact = await get_contact(db, user, contact_id)
+    if contact is None:
+        return None
     db.delete(contact)
     db.commit()
 
@@ -79,6 +144,17 @@ async def search_contacts(
         user: User,
         search_query: str
 ) -> List[Contact]:
+    """
+    Search for contacts for a specific user
+
+    Args:
+        db (Session): The database session.
+        user (User): The user to search contacts for.
+        search_query (str): The search query to use. Will search in first_name, last_name, email, and phone.
+
+    Returns:
+        List[Contact]: A list of contacts that match the search query.
+    """
     return db.query(Contact).filter(
         and_(
             Contact.user_id == user.id,
@@ -92,12 +168,24 @@ async def search_contacts(
     ).all()
 
 
-async def get_7_days_birthday_contacts(
+async def get_birthdays_contacts(
         db: Session,
-        user: User
+        user: User,
+        days: int = 7
 ) -> List[Contact]:
+    """
+    Get a list of contacts that have birthdays within the next `days` days.
+
+    Args:
+        db (Session): The database session.
+        user (User): The user to get contacts for.
+        days (int, optional): The number of days to look ahead for birthdays. Defaults to 7.
+
+    Returns:
+        List[Contact]: A list of contacts that have birthdays within the next `days` days.
+    """
     today = datetime.today()
-    seven_days_later = today + timedelta(days=7)
+    seven_days_later = today + timedelta(days=days)
 
     # Common filters
     user_filter = Contact.user_id == user.id
